+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Fix an issue where an empty file in `FileCache` could cause a parsing error. (#181)
+- Support caching for `POST` and other HTTP methods. (#183)
 
 ## 0.0.22 (31th January, 2024)